SER (Distributed Network Architecture) service chain
Service, referred to as SER, Chinese name service chain, established in a centralized service system. So that everyone and organizations and developers can participate in the construction of the service system framework. The service chain system is based on the blockchain technology such as information technology, business alliance, consumer union and system engineering, and the open source ecosystem. Open source sharing, and consensus mechanisms to encrypt the currency, decentralizing the service-related aspects such as the hospitality industry, hotel industry, insurance industry, service industry, etc. can be used to radically improve the ecological infrastructure and organize it in an ecologically- Get together to form a complete consumer service network.